Move files out of the root dom/ folder
