+++ index.html
@@ -399,11 +399,11 @@
       }
     </script>
     <script type="module" src="js/main.js"></script>
-    <script type="module" src="js/modals/welcome.js"></script>
-    <script type="module" src="js/modals/input.js"></script>
-    <script type="module" src="js/information.js"></script>
-    <script type="module" src="js/views/views.js"></script>
-    <script type="module" src="js/current-file.js"></script>
-    <script type="module" src="js/switch-deploy.js"></script>
+    <script type="module" src="js/dom/modals/welcome.js"></script>
+    <script type="module" src="js/dom/modals/input.js"></script>
+    <script type="module" src="js/dom/modals/information.js"></script>
+    <script type="module" src="js/dom/views/views.js"></script>
+    <script type="module" src="js/dom/state/current-file.js"></script>
+    <script type="module" src="js/dom/toggle/switch-deploy.js"></script>
   </body>
 </html>
